dj004/requirement.txt
anyio==3.6.2
asgiref==3.5.2
async-timeout==4.0.2
backcall==0.2.0
click==8.1.3
colorama==0.4.5
decorator==5.1.1
Deprecated==1.2.13
Django==3.2.15
django-filter==22.1
djangorestframework==3.14.0
fastapi==0.87.0
Flask==2.2.2
h11==0.14.0
idna==3.4
importlib-metadata==4.12.0
ipython==7.34.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
numpy==1.21.6
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
pip==22.3.1
prompt-toolkit==3.0.31
pydantic==1.10.2
Pygments==2.13.0
PyMySQL==1.0.2
pyparsing==3.0.9
pytz==2022.2.1
redis==4.3.4
setuptools==65.3.0
sniffio==1.3.0
sqlparse==0.4.2
starlette==0.21.0
traitlets==5.4.0
typing_extensions==4.3.0
uvicorn==0.20.0
wcwidth==0.2.5
Werkzeug==2.2.2
wheel==0.37.1
wrapt==1.14.1
zipp==3.8.1
